improve: Display "raw" or "processed" in files table, allow "completed" files to be reprocessed
